requirements.txt
flask
keras
tensorflow-cpu
pillow
pandas
gunicorn
